+++ packages/expo-file-system/android/src/main/java/expo/modules/filesystem/FileSystemUploadTask.kt
@@ -5,6 +5,7 @@ import expo.modules.filesystem.unifiedfile.UnifiedFileInterface
 import expo.modules.kotlin.types.Enumerable
 import expo.modules.kotlin.records.Field
 import expo.modules.kotlin.records.Record
+import expo.modules.kotlin.services.FilePermissionService
 import expo.modules.kotlin.sharedobjects.SharedObject
 import expo.modules.kotlin.types.OptimizedRecord
 import kotlinx.coroutines.suspendCancellableCoroutine
@@ -83,6 +84,7 @@ class FileSystemUploadTask : SharedObject() {
   private val progressThrottleInterval = 100.milliseconds
 
   suspend fun start(url: String, file: FileSystemFile, options: UploadTaskOptions): UploadTaskResult {
+    file.validatePermission(FilePermissionService.Permission.READ)
     cancelled = false
     val request = buildUploadRequest(url, file, options)
     return executeUploadRequest(request)